Do a bit more cleanup
