--- libs/shared/ui/chart/src/lib/chart-options.ts
@@ -1,7 +1,0 @@
-export interface ChartOptions {
-  title: string;
-  type: string;
-  data: any;
-  columnNames: string[];
-  options: any;
-}
