+++ libs/shared/ui/chart/src/lib/chart/chart.component.spec.ts
@@ -23,7 +23,7 @@ describe('ChartComponent', () => {
       type: 'LineChart',
       data: [],
       columnNames: ['period', 'close'],
-      options: {title: `Stock price`, width: '600', height: '400'}
+      options: { title: `Stock price`, width: '600', height: '400' }
     };
     component.data = [];
