+++ libs/stocks/feature-shell/src/lib/stocks/stocks.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { ChartOptions } from '@coding-challenge/shared/ui/chart';
 import { PriceQueryFacade } from '@coding-challenge/stocks/data-access-price-query';
+import { ChartOptions } from '../../../../../shared/ui/chart/src/lib/chart/chart.component';
 import { StockPickerSelection } from '../stock-picker/stock-picker.component';
 
 @Component({
